ip_tracking/requirements.txt
django-ipware>=3.0.0
django-ratelimit>=4.0.0
requests>=2.25.1
